Initial commit
cc @jperl @derekcaneja Deep link package, note the helper functions for
generating links.

On iOS the banner will show at the top of the screen if used
`iosBanner` for android theres an option for adding a fallback url eg.
to the store.

Theres a small qa app for testing / demo.